requirements.txt
fastapi
uvicorn[standard]
python-multipart  
celery
redis
sqlalchemy
psycopg2-binary
pydantic-settings
pdfplumber
python-docx
pytesseract
pdf2image
Pillow
google-generativeai
pytest
httpx
